Simplificado actuno_poruno

+++ .bash_aliases_debian.sh
@@ -59,17 +59,16 @@ sudo bash -c "aptitude search -F '%p' --disable-columns '~U'| grep -v -e ^lib[a-
 actuno_por_uno(){
 echo "################ #################### ########################
 Actualiza uno por uno."
-sudo bash -c "for i in `aptitude search '~U' -F %p`; do
+sudo bash -c 'for i in `aptitude search \'~U' -F %p`; do
+if echo ${i}|grep -v -e ^lib[a-q] -e ^lib[s-z] -e ^wine -e python -e plasma -e ruby -e ^glib -e common -e data -e ^gir1. -e ^libr[a-d] -e ^libr[f-z] -e ^libre[a-n] -e ^libre[p-z] -e ^mono > /dev/null; then
+echo "Analizando $i"
 killall apt-get apt-mark
-apt-get install --no-remove -q=2 --allow-unauthenticated $i && \
-apt-mark auto $i
-done
-aptitude"
+apt-get install --no-remove -q=2 --allow-unauthenticated ${i} && apt-mark auto ${i}
+fi
+done'
+aptitude
 read -p "Enter para continuar con posibilidad de preguntar si borrar algún paquete" a
-sudo bash -c "aptitude search -F '%p' --disable-columns '~U'| grep -v -e ^lib[a-q] -e ^lib[s-z] -e ^wine -e python -e plasma -e ruby -e ^glib -e common -e data -e ^gir1. -e ^libr[a-d] -e ^libr[f-z] -e ^libre[a-n] -e ^libre[p-z]|xargs -l1 apt-get install --allow-unauthenticated" 
-#paplay /usr/share/sounds/KDE-Im-Nudge.ogg
-#apt-get upgrade -s |grep 'Inst '| cut -d' ' -f2| grep -v -e ^lib[a-q] -e ^lib[s-z] -e ^libr[a-d] -e ^libr[f-z] -e ^libre[a-n] -e ^libre[p-z] -e ^uno -e ^ure -e ^wine -e python -e plasma -e ruby -e ^glib -e common -e data -e ^gir1. -e python |xargs -l1 apt-get install --no-remove -y"
-##aptitude install --safe-resolver --allow-new-installs --allow-untrusted -y
+aptitude search -F '%p' --disable-columns '~U'|xargs -l1 sudo apt-get install --allow-unauthenticated
 }
 
 #it(){
